+++ hw/pci.h
@@ -20,8 +20,8 @@ typedef void PCIMapIORegionFunc(PCIDevice *pci_dev, int region_num,
 #define PCI_ADDRESS_SPACE_MEM_PREFETCH	0x08
 
 typedef struct PCIIORegion {
-    uint32_t addr; /* current PCI mapping address. -1 means not mapped */
-    uint32_t size;
+    unsigned long addr; /* current PCI mapping address. -1 means not mapped */
+    unsigned long size;
     uint8_t type;
     PCIMapIORegionFunc *map_func;
 } PCIIORegion;
@@ -71,7 +71,7 @@ PCIDevice *pci_register_device(PCIBus *bus, const char *name,
                                PCIConfigWriteFunc *config_write);
 
 void pci_register_io_region(PCIDevice *pci_dev, int region_num,
-                            uint32_t size, int type,
+                            unsigned long size, int type,
                             PCIMapIORegionFunc *map_func);
 
 uint32_t pci_default_read_config(PCIDevice *d,
@@ -103,7 +103,7 @@ void *lsi_scsi_init(PCIBus *bus, int devfn);
 
 /* vmware_vga.c */
 void pci_vmsvga_init(PCIBus *bus, DisplayState *ds, uint8_t *vga_ram_base,
-                     unsigned long vga_ram_offset, int vga_ram_size);
+                     unsigned long vga_ram_offset, unsigned long vga_ram_size);
 
 /* usb-uhci.c */
 void usb_uhci_piix3_init(PCIBus *bus, int devfn);
